backend/requirements.txt
google-genai
fastapi
uvicorn[standard]
python-dotenv
